Add an option for TLS server certs
Allow configuring the trust store used to verify connections to Redis.
This is useful when working with something like GCP Memorystore for
Redis ([1]), which will issue a self-signed cert for managed Redis
instances. With this option, I can use Terraform to create the managed
Redis instance, get the server cert that was generated, and render it
into the Caddy configuration file.

[1]: https://cloud.google.com/memorystore/docs/redis